COG-654: Office lists are ordered by office name
