+++ src/test/resources/fixtures/idm/admin-offices/all-offices.json
@@ -1,27 +1,27 @@
 [
   {
-    "office_id": "JzoGSkE05I",
+    "office_id": "JzoGSkE07I",
     "county_id": 1124,
     "county_name": "Yolo",
-    "office_name": "Tolo Office"
+    "office_name": "11 Tolo Office"
   },
   {
     "office_id": "JzoGSkE06I",
     "county_id": 1124,
     "county_name": "Yolo",
-    "office_name": "Tolo Office"
-  },
-  {
-    "office_id": "JzoGSkE07I",
-    "county_id": 1124,
-    "county_name": "Yolo",
-    "office_name": "Tolo Office"
+    "office_name": "2 Tolo Office"
   },
   {
     "office_id": "BgUwJ2Uaat",
     "county_id": 1084,
     "county_name": "Lake",
     "office_name": "Stan's CWDS Office"
+  },
+  {
+    "office_id": "JzoGSkE05I",
+    "county_id": 1124,
+    "county_name": "Yolo",
+    "office_name": "Tolo Office"
   },
   {
     "office_id": "TG7O51q0Ki",
