+++ src/test/resources/liquibase/cms-data.xml
@@ -87,7 +87,7 @@
       <column name="COMNT_DSC" value="                                                                                                                        "/>
       <column name="AGENCY_NM" value="Yolo Agency                                  "/>
       <column name="DPT_DIV_NM" value="Yolo Department                              "/>
-      <column name="CWS_OFF_NM" value="Tolo Office                                  "/>
+      <column name="CWS_OFF_NM" value="2 Tolo Office                                  "/>
       <column name="CNTY_SPFCD" value="57"/>
       <column name="AGCY_CD_NO" valueNumeric="0"/>
       <column name="LOC_CNTY" valueNumeric="0"/>
@@ -151,7 +151,7 @@
       <column name="COMNT_DSC" value="                                                                                                                        "/>
       <column name="AGENCY_NM" value="Yolo Agency                                  "/>
       <column name="DPT_DIV_NM" value="Yolo Department                              "/>
-      <column name="CWS_OFF_NM" value="Tolo Office                                  "/>
+      <column name="CWS_OFF_NM" value="11 Tolo Office                                  "/>
       <column name="CNTY_SPFCD" value="57"/>
       <column name="AGCY_CD_NO" valueNumeric="0"/>
       <column name="LOC_CNTY" valueNumeric="0"/>
